Split MFIIndicator into several files.
