+++ ts/Stock/Indicators/MFI/MFIPoint.d.ts
@@ -1,0 +1,33 @@
+/* *
+ *
+ *  License: www.highcharts.com/license
+ *
+ *  !!!!!!! SOURCE GETS TRANSPILED BY TYPESCRIPT. EDIT TS FILE ONLY. !!!!!!!
+ * */
+
+/* *
+ *
+ *  Imports
+ *
+ * */
+
+import MFIIndicator from './MFIIndicator';
+import type SMAPoint from '../SMA/SMAPoint';
+
+/* *
+*
+*  Class
+*
+* */
+
+declare class MFIPoint extends SMAPoint {
+    public series: MFIIndicator;
+}
+
+/* *
+*
+*  Default Export
+*
+* */
+
+export default MFIPoint;
